requirements.txt
dj-database-url==0.4.1
Django==1.10.2
django-cors-headers==1.3.1
django-suit==0.2.22
djangorestframework==3.5.3
drf-nested-routers==0.11.1
gunicorn==19.6.0
Pillow==3.4.2
psycopg2==2.6.2
pytz==2016.10
requests==2.12.4
whitenoise==3.2.2
